requirements.txt
# Voice Bot Dependencies

# Telephony
twilio>=8.0.0

# Speech Services (Deepgram for STT + TTS)
websockets>=12.0

# LLM
anthropic>=0.25.0

# Configuration
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.27.0
requests>=2.31.0

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
